+++ include/nstd/gl.h
@@ -3,6 +3,7 @@
 #include "core/def.h"
 #include "core/slice.h"
 #include "gui.h"
+#include "str.h"
 #ifdef __cplusplus
 extern "C"
 {
@@ -82,7 +83,7 @@ typedef enum
 /// Contains information on a device.
 typedef struct
 {
-    char *name;
+    NSTDString name;
     NSTDUSize vendor;
     NSTDUSize device;
     NSTDGLDeviceType device_type;